Add script for events audit
There are multiple ways how code in Supervisor may trigger data collection to MixPanel.
This script was used to review all possible triggers and summarize then in a CSV file.
It does not pretend to automate the audit - it's still a manuall work, but this script
helps to verify and organize results.

Change-type: patch
Signed-off-by: Roman Mazur <roman@balena.io>