Set a mock l4t version for use with optional containers
Before this change a mixed device fleet with a subset using l4t would
not be able to have containers which only run on the l4t based boards.
We add a mock l4t version so the contracts will always be denied on
non-l4t boards, but the release itself won't fail to validate.

Change-type: patch
Signed-off-by: Cameron Diver <cameron@balena.io>