Add a noinit option to sync-debug
The noinit option will avoid doing a complete sync to the supervisor
on startup. This is useful when there has only been code changes, and no
modules etc have been added. It can be helpful when running on a network
with no external internet access, as everything needed is already inside
the container.

To aid this we also move to yargs, rather than try to do all of the
parsing ourselves.

Change-type: minor
Signed-off-by: Cameron Diver <cameron@balena.io>