Migrate to per-service apiSecrets
We have a api key per service, and one more key which has all
permissions (used for the cloud). When a key is used for the
supervisor api, we pull in the scopes for endpoints to respect
as part of their normal operation.

The scopes currently supported are `app` and `all-apps` but this
will soon be extended.

Change-type: minor
Signed-off-by: Cameron Diver <cameron@balena.io>