Support app uuids and release version on target state
The new v3 device state endpoint uses app uuids as the identifier for the
app and adds releaseVersion and isHost as optional attributes of an app.
The v3 target state endpoint is also a tri-app target state, including
the supervisor and host app in addition to the user app.

`releaseVersion` will eventually become the main identifier of a release
and `isHost` will allow to identify the host app on the target state which
will indicate the supervisor that is an app allowed to install services
and overlays on the root partition.

For now this change still uses environment dependent ids (serviceId,
imageId, releaseId) for comparing
current and target state and for reporting current state of releases and
images back to the cloud, but that should be changing soon.

We store the uuid and releaseVersion on the individual docker components so that when
building a current state, we can match them with the target state.

This is a breaking change

Change-type: major
Depends-on: https://github.com/balena-io/open-balena-api/pull/504
Depends-on: #1676