Bootstrap supervisor configuration on container start
This change goes in line with making the supervisor more like an app. This
reduces the number of parameters that need to be given to start the
supervisor container, moving it closer to something that could be
started by the supervisor itself. This should allow to greatly simplify
the supervisor [start script](https://github.com/balena-os/meta-balena/blob/master/meta-balena-common/recipes-containers/balena-supervisor/balena-supervisor/start-balena-supervisor)

This is now the minimal startup configuration for the container

```
balena run --privileged --name balena_supervisor \
	--restart=always \
	--net=host \
	--cidenv=SUPERVISOR_CONTAINER_ID \
	--mount type=bind,source=/var/run/balena-engine.sock,target=/var/run/balena-engine.sock \
	--mount type=bind,source=/,target=/mnt/root \
	-e DOCKER_SOCKET=/var/run/balena-engine.sock \
	"${SUPERVISOR_IMAGE}:${SUPERVISOR_TAG}"
```

Change-type: patch