Remove releaseId from container name
The release id is not really necessary for reporting the current state
since v3 (it has been replaced by `releaseUuid`) and is only kept for
backwards compatibility for some supervisor API endpoints.

Until now, the `releaseId` was part o the container name, leading to
longer names than desired. This change removes the value from the
container name and modifies queries to look for the value in the image
database.

Change-type: minor
Relates-to: #2077